Added Pug, Autoprefixer, and Sanitize.css
* Set up fuse such that it would take a index.pug file in /src and use
it as a template for the webindex plugin. I'm not entirely happy with it
and I will likely change it later. Naturally index.html was replaced
with index.pug

* Set up autoprefixer and Santize.css, added the postcss plugin to
fuse.js and added antoher cssplugin as I realised the other one would
only work on sass files.

* changed double quotes to single quotes and added some extra spacing to
fuse.js for legibility